Split parser methods into its own DAO class
